app/requirements.txt
# app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
pydantic>=1.8.0
python-dotenv>=0.19.0
python-telegram-bot==13.15